avltree . insert

+++ include/top/core/stddef.h
@@ -6,8 +6,8 @@
 extern "C" {
 #endif
 
-#define top_offset_of(type,member) ((size_t)(&((type*)0)->member))
-#define top_container_of(ptr,type,member) ((type*)((const char*)ptr - top_offset_of(type,member)))
+#define top_offset_of(type,member) ((unsigned long)(&((type*)0)->member))
+#define top_container_of(ptr,type,member) ((type*)((const char*)(ptr) - top_offset_of(type,member)))
 
 #ifdef __cplusplus
 }
